returned the white buildings from previous versions
added green and blue graph streams to all vision sensor graphs, so that green and blue values can be visualized also

added new constraint to the vision detection, in order to avoid red vision sensor to be triggered by white buildings, now the green and blue values must be at most half of the red value to be considered a potential survivor. if not, it is most likely a white building in the scene that it could have detected.